#40 early 3.4.0-beta.1: refactor to follow one main common activity and bump environment
Also bring package private factory for YoutubePlayerView to bypass required YoutubeBaseActivity
