+++ uxkit-common/src/main/java/com/voxeet/uxkit/common/service/SDKBinder.java
@@ -1,4 +1,4 @@
-package com.voxeet.uxkit.service;
+package com.voxeet.uxkit.common.service;
 
 import android.os.Binder;
