--- uxkit-incoming-call/src/main/java/com/voxeet/uxkit/incoming/factory/IVoxeetActivity.java
@@ -1,5 +1,0 @@
-package com.voxeet.uxkit.incoming.factory;
-
-//TODO this should be something set in the SDK to remove this dependency as mandatory
-public interface IVoxeetActivity {
-}
