+++ uxkit-youtube/src/main/res/layout/voxeet_youtube.xml
@@ -1,11 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/voxeet_youtube_parent"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <com.google.android.youtube.player.YouTubePlayerView
-        android:id="@+id/voxeet_youtube"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-</FrameLayout>
+    android:orientation="vertical" />
